feat: small changes
