+++ src/app/pokemon/[id]/details/components/Evolution.tsx
@@ -1,0 +1,1 @@
+export function Evolution() {}
