+++ src/app/pokemon/[id]/details/components/Status.tsx
