+++ src/app/pokemon/[id]/details/page.tsx
@@ -1,21 +1,6 @@
 import { PokemonContainer } from '../../components/PokemonContainer';
-import { Species } from '../components/Species';
-import { Status } from '../components/Status';
-
-const species = {
-  description:
-    ' Lorem ipsum dolor, sit amet consectetur adipisicing elit. Sunt iste voluptates doloribus ipsa vel impedit suscipit, animi quod!',
-  height: '2\'4\" (0.71 m)',
-  weight: '10 lbs (4.5 kg)',
-};
-
-const habilities = [
-  {
-    name: 'Hability 1',
-    description:
-      'Lorem ipsum dolor, sit amet consectetur adipisicing elit. Sunt iste voluptates doloribus ipsa vel impedit suscipit, animi quod!',
-  },
-];
+import { Species } from './components/Species';
+import { Status } from './components/Status';
 
 export default function Details() {
   return (
@@ -26,6 +11,7 @@ export default function Details() {
       <PokemonContainer title="Status">
         <Status />
       </PokemonContainer>
+      <PokemonContainer title="Evolution">1</PokemonContainer>
     </section>
   );
 }
